Move handlers, return ServeRoot HandlerFunc
